added css screen change and timer stuff
The start screen can progress to the game screen now, and starts a
countdown. Pause and resume buttons are implemented that stop and resume
the timer. We can use those functions in the JS to do their respective
actions for the whole game. Theyre placed in an ugly manner atm, that's
all.